Create README.md
This is a Practice project. the framework has been developed using POM, BDD Cucumber Java using gherkin syntax for test cases and Junit for running the test. It is also built with Maven as this help me to automatically download the dependencies.

The signup cucumber feature file has a corresponding java classes, one for the step definitions, the page class is where I define all my methods and also declare all the elements I need for that particular class. the same with the other two feature files.

I have an Util package where I have created the base class for triggering the test execution.
The @Test Annotations on the feature files is for cucumber to identify the files I want to test.